return image in uint8 format

+++ astrovision/plot/plot_utils.py
@@ -71,7 +71,7 @@ def make_mosaic(
     # Create SatelliteImage
     # TODO: dep and date if all images have same dep and date
     mosaic_image = SatelliteImage(
-        array=mosaic,
+        array=mosaic.astype("uint8"),
         crs=satellite_images[0].crs,
         bounds=bounds,
         transform=out_transform,
